Re-organize user creation and show new user password window
