+++ src/components/Admin/UserPage.tsx
@@ -1,33 +1,42 @@
 import React, { FC } from 'react';
-import { Button, Col, Row } from 'react-bootstrap';
+import { Col, Row } from 'react-bootstrap';
 import { Link, Route, Switch, useRouteMatch } from 'react-router-dom';
 import { useUsersQuery } from '../../generated/graphql';
-import { defaultUser, UserModel } from '../../models/User';
+import { UserModel } from '../../models/User';
+import Button from '../core/Button';
 import Loading from '../core/Loading';
-import User from './User';
-import UserForm, { EditMode } from './UserForm';
+import UserEdit from './UserEdit';
+import { EditMode } from './UserForm';
 import UserList from './UserList';
 
+// type UserPageProps = PageProps;
+
+// export const UserPage: FC<UserPageProps> = ({ paths }) => {
 export const UserPage: FC = () => {
-  const { path } = useRouteMatch();
+  const { path, url } = useRouteMatch();
   const { data, loading } = useUsersQuery();
+
+  // const currentPaths = useMemo(() => [...paths, { value: url, name: 'users', real: true }], [paths]);
+  // useUpdateNavigation({ currentPaths });
+
   const users = (data?.users || []) as UserModel[];
 
   const title = 'Users';
 
   if (loading) return <Loading />;
+
   return (
     <>
       <Row className="justify-content-end">
-        <Col>
-          <h2 style={{ textAlign: 'center' }}>{title}</h2>
-        </Col>
         <Col sm={1}>
           <Link to={`${path}/new`}>
-            <Button>
-              <span>New</span>
+            <Button variant="primary" small>
+              New
             </Button>
           </Link>
+        </Col>
+        <Col>
+          <h2 style={{ textAlign: 'center' }}>{title}</h2>
         </Col>
       </Row>
       <hr />
@@ -38,13 +47,13 @@ export const UserPage: FC = () => {
               <UserList users={users} />
             </Route>
             <Route path={`${path}/new`}>
-              <UserForm user={defaultUser} editMode={EditMode.new} title={'User creation'} />
+              <UserEdit editMode={EditMode.new} title={'User creation'} />
             </Route>
             <Route exact path={`${path}/:userId/edit`}>
-              <User mode={EditMode.edit} />
+              <UserEdit editMode={EditMode.edit} />
             </Route>
             <Route exact path={`${path}/:userId`}>
-              <User mode={EditMode.readOnly} />
+              <UserEdit editMode={EditMode.readOnly} />
             </Route>
           </Switch>
         </Col>
